Dynamic import to 'react-color
